separating the template into components and adding it in the config
