+++ src/Pages/Home.tsx
@@ -1,0 +1,48 @@
+import * as React from 'react';
+import Button from '@mui/material/Button';
+import Stack from '@mui/material/Stack';
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+
+export default function Home() {
+  return (
+    <>
+      <main>
+        {/* Hero unit */}
+        <Box
+          sx={{
+            bgcolor: 'background.paper',
+            pt: 8,
+            pb: 6,
+          }}
+        >
+          <Container maxWidth="md">
+            <Typography
+              component="h1"
+              variant="h2"
+              align="center"
+              color="text.primary"
+              gutterBottom
+            >
+              Thumbnail Generator
+            </Typography>
+            <Typography variant="h5" align="center" color="text.secondary" paragraph>
+                
+            </Typography>
+            <Stack
+              sx={{ pt: 4 }}
+              direction="row"
+              spacing={2}
+              justifyContent="center"
+            >
+              <Button variant="contained">Crop your image</Button>
+              <Button variant="outlined">Resize your image</Button>
+            </Stack>
+          </Container>
+        </Box>
+          
+      </main>
+    </>
+  );
+}
